MWPW-151936 Aside Tiger Team Enhancements
