+++ test/blocks/aside/mocks/split.html
@@ -1,0 +1,114 @@
+<div class="aside split medium" id="test-default">
+  <div>
+    <div data-valign="middle">#1e1e1e</div>
+  </div>
+  <div>
+    <div data-valign="middle">
+      <p>
+        <picture>
+          <img loading="lazy" alt="" src="" width="166" height="160">
+        </picture>
+      </p>
+      <p>Detail</p>
+      <h3 id="heading--aside-split-medium">Heading – Aside (split, medium)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <ul>
+        <li>
+          <picture>
+            <img loading="lazy" alt="https://milo.adobe.com | Image" src="" width="154" height="150">
+          </picture> <a href="https://www.adobe.com/">Adobe Photoshop 1</a>
+        </li>
+        <li>
+          <picture>
+            <img loading="lazy" alt="" src="" width="154" height="150">
+          </picture><a href="https://www.adobe.com/">Adobe Photoshop 2</a>
+        </li>
+      </ul>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <strong><a href="https://www.adobe.com/">Learn More</a></strong> <a href="https://www.adobe.com/">Text Link</a></p>
+      <p>Body CTA supplemental text (optional)</p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="2320" height="840">
+      </picture>
+    </div>
+  </div>
+</div>
+<div class="aside split l-heading m-body l-detail" id="test-text-overrides">
+  <div>
+    <div data-valign="middle">#1e1e1e</div>
+  </div>
+  <div>
+    <div data-valign="middle">
+      <p>
+        <picture>
+          <img loading="lazy" alt="" src="" width="166" height="160">
+        </picture>
+      </p>
+      <p>Detail</p>
+      <h3 id="heading--aside-split-l-heading-m-body-l-detail">Heading – Aside (split, l-heading, m-body, l-detail)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <strong><a href="https://www.adobe.com/">Learn More</a></strong> <a href="https://www.adobe.com/">Text Link</a></p>
+      <p>Body CTA supplemental text (optional)</p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="2320" height="840">
+      </picture>
+    </div>
+  </div>
+</div>
+<div class="aside split l-title" id="test-title">
+  <div>
+    <div data-valign="middle">
+      <p>Title</p>
+      <h3 id="heading--aside-split-l-title">Heading – Aside (split, l-title)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <strong><a href="https://www.adobe.com/">Learn More</a></strong> <a href="https://www.adobe.com/">Text Link</a></p>
+      <p>Body CTA supplemental text (optional)</p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="2320" height="840">
+      </picture>
+    </div>
+  </div>
+</div>
+<div class="aside split l-avatar" id="test-avatar">
+  <div>
+    <div data-valign="middle">
+      <p>
+        <picture>
+          <img loading="lazy" alt="" src="" width="166" height="160">
+        </picture>
+      </p>
+      <p>Detail</p>
+      <h3 id="heading--aside-split-l-avatar">Heading – Aside (split, l-avatar)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <strong><a href="https://www.adobe.com/">Learn More</a></strong> <a href="https://www.adobe.com/">Text Link</a></p>
+      <p>Body CTA supplemental text (optional)</p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="2320" height="840">
+      </picture>
+    </div>
+  </div>
+</div>
+<div class="aside split l-lockup" id="test-lockup">
+  <div>
+    <div data-valign="middle">
+      <p><a href=""><picture><img loading="lazy" src=""></picture></a> Adobe</p>
+      <p>Detail</p>
+      <h3 id="heading--aside-split-l-avatar">Heading – Aside (split, l-avatar)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <strong><a href="https://www.adobe.com/">Learn More</a></strong> <a href="https://www.adobe.com/">Text Link</a></p>
+      <p>Body CTA supplemental text (optional)</p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="2320" height="840">
+      </picture>
+    </div>
+  </div>
+</div>
