+++ test/blocks/aside/mocks/standard.html
@@ -1,0 +1,97 @@
+<div class="aside" id="test-default">
+  <div>
+    <div data-valign="middle">#eeeeee</div>
+  </div>
+  <div>
+    <div data-valign="middle">
+      <p>
+        <picture>
+          <img loading="lazy" alt="" src="" width="154" height="150">
+        </picture>
+      </p>
+      <p>Detail</p>
+      <h3 id="heading--aside">Heading – Aside</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <a href="https://www.adobe.com/">Text Link</a></p>
+      <p>Body CTA supplemental text (optional)</p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="500" height="375">
+      </picture>
+    </div>
+  </div>
+</div>
+<div class="aside" id="test-default-2">
+  <div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="750" height="375">
+      </picture>
+    </div>
+  </div>
+  <div>
+    <div data-valign="middle">
+      <p><strong>Detail</strong></p>
+      <h3 id="heading--aside">Heading – Aside</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <a href="https://www.adobe.com/">Text Link</a></p>
+    </div>
+  </div>
+</div>
+<div class="aside center l-heading m-body l-detail" id="test-text-overrides">
+  <div>
+    <div data-valign="middle">
+      <p>Detail</p>
+      <h3 id="heading--aside-center-l-heading-m-body-l-detail">Heading – Aside (center, l-heading, m-body, l-detail)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <a href="https://www.adobe.com/">Text Link</a></p>
+    </div>
+  </div>
+</div>
+<div class="aside center l-title" id="test-title">
+  <div>
+    <div data-valign="middle">
+      <p>Title</p>
+      <h3 id="heading--aside-center-l-title">Heading – Aside (center, l-title)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <a href="https://www.adobe.com/">Text Link</a></p>
+    </div>
+  </div>
+</div>
+<div class="aside l-avatar" id="test-avatar">
+  <div>
+    <div data-valign="middle">
+      <p>
+        <picture>
+          <img loading="lazy" alt="" src="" width="86" height="86">
+        </picture>
+      </p>
+      <p>Detail</p>
+      <h3 id="heading--aside-l-avatar">Heading – Aside (l-avatar)</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <a href="https://www.adobe.com/">Text Link</a></p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="500" height="375">
+      </picture>
+    </div>
+  </div>
+</div>
+<div class="aside m-lockup" id="test-lockup">
+  <div>
+    <div data-valign="middle">
+      <p><a href=""><picture><img loading="lazy" src=""></picture></a> Adobe</p>
+      <p>Detail</p>
+      <h3 id="heading--aside">Heading – Aside</h3>
+      <p>Body Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore <a href="https://www.adobe.com/">magna</a>.</p>
+      <p><em><a href="https://www.adobe.com/">Learn More</a></em> <a href="https://www.adobe.com/">Text Link</a></p>
+    </div>
+    <div data-valign="middle">
+      <picture>
+        <img loading="lazy" alt="" src="" width="500" height="375">
+      </picture>
+    </div>
+  </div>
+</div>
